requirements.txt
wyoming==1.5.3
coqui-tts==0.26.0
coqui-tts[server]==0.26.0
deepspeed==0.16.7